Moving log_blank_file and exit_if_nonlocal into function_bag
